+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Version 1.8.0 (not yet released)
 
+
+
+# Version 1.7.1
+
 * Fixed issue with cloning data processors.
 * Added Manage Case Link field.
 * Fixed issue with dashlet opening in full screen.